+++ .gitignore
@@ -28,6 +28,7 @@ cranelift/isle/veri/veri_engine/test_output
 crates/explorer/node_modules
 tests/all/pulley_provenance_test.cwasm
 tests/all/pulley_provenance_test_component.cwasm
+tests/all/pulley_provenance_test_async_component.cwasm
 /artifacts
 testcase*.wat
 testcase*.wasm
