+++ crates/wasmtime/Cargo.toml
@@ -195,8 +195,10 @@ async = [
   "dep:wasmtime-fiber",
   "dep:async-trait",
   "dep:trait-variant",
+  "dep:futures",
   "wasmtime-component-macro?/async",
   "runtime",
+  "futures/std",  
 ]
 
 # Enables support for the pooling instance allocation strategy
