+++ crates/wasmtime/src/runtime/component/linker.rs
@@ -7,6 +7,8 @@ use crate::component::types;
 use crate::component::{
     Component, ComponentNamedList, Instance, InstancePre, Lift, Lower, ResourceType, Val,
 };
+#[cfg(feature = "async")]
+use crate::fiber::AsyncCx;
 use crate::hash_map::HashMap;
 use crate::prelude::*;
 use crate::{AsContextMut, Engine, Module, StoreContextMut};
@@ -443,18 +445,9 @@ impl<T: 'static> LinkerInstance<'_, T> {
         );
 
         let ff = move |mut store: StoreContextMut<'_, T>, params: Params| -> Result<Return> {
-            #[cfg(feature = "component-model-async")]
-            {
-                let async_cx = crate::component::concurrent::AsyncCx::new(&mut store.0);
-                let mut future = Pin::from(f(store.as_context_mut(), params));
-                unsafe { async_cx.block_on(future.as_mut()) }?
-            }
-            #[cfg(not(feature = "component-model-async"))]
-            {
-                let async_cx = store.as_context_mut().0.async_cx().expect("async cx");
-                let future = f(store.as_context_mut(), params);
-                unsafe { async_cx.block_on(Pin::from(future)) }?
-            }
+            let async_cx = AsyncCx::new(&mut store.0);
+            let mut future = Pin::from(f(store.as_context_mut(), params));
+            async_cx.block_on(future.as_mut())?
         };
         self.func_wrap(name, ff)
     }
@@ -618,18 +611,9 @@ impl<T: 'static> LinkerInstance<'_, T> {
             "cannot use `func_new_async` without enabling async support in the config"
         );
         let ff = move |mut store: StoreContextMut<'_, T>, params: &[Val], results: &mut [Val]| {
-            #[cfg(feature = "component-model-async")]
-            {
-                let async_cx = crate::component::concurrent::AsyncCx::new(&mut store.0);
-                let mut future = Pin::from(f(store.as_context_mut(), params, results));
-                unsafe { async_cx.block_on(future.as_mut()) }?
-            }
-            #[cfg(not(feature = "component-model-async"))]
-            {
-                let async_cx = store.as_context_mut().0.async_cx().expect("async cx");
-                let future = f(store.as_context_mut(), params, results);
-                unsafe { async_cx.block_on(Pin::from(future)) }?
-            }
+            let async_cx = AsyncCx::new(&mut store.0);
+            let mut future = Pin::from(f(store.as_context_mut(), params, results));
+            async_cx.block_on(future.as_mut())?
         };
         self.func_new(name, ff)
     }
@@ -727,22 +711,9 @@ impl<T: 'static> LinkerInstance<'_, T> {
         let dtor = Arc::new(crate::func::HostFunc::wrap_inner(
             &self.engine,
             move |mut cx: crate::Caller<'_, T>, (param,): (u32,)| {
-                #[cfg(feature = "component-model-async")]
-                {
-                    let async_cx =
-                        crate::component::concurrent::AsyncCx::new(&mut cx.as_context_mut().0);
-                    let mut future = Pin::from(dtor(cx.as_context_mut(), param));
-                    unsafe { async_cx.block_on(future.as_mut()) }?
-                }
-                #[cfg(not(feature = "component-model-async"))]
-                {
-                    let async_cx = cx.as_context_mut().0.async_cx().expect("async cx");
-                    let future = dtor(cx.as_context_mut(), param);
-                    match unsafe { async_cx.block_on(Pin::from(future)) } {
-                        Ok(Ok(())) => Ok(()),
-                        Ok(Err(trap)) | Err(trap) => Err(trap),
-                    }
-                }
+                let async_cx = AsyncCx::new(&mut cx.as_context_mut().0);
+                let mut future = Pin::from(dtor(cx.as_context_mut(), param));
+                async_cx.block_on(future.as_mut())?
             },
         ));
         self.insert(name, Definition::Resource(ty, dtor))?;
