+++ crates/wasmtime/src/runtime/externals/table.rs
@@ -104,25 +104,14 @@ impl Table {
         ty: TableType,
         init: Ref,
     ) -> Result<Table> {
-        let store = store.as_context_mut();
+        let mut store = store.as_context_mut();
         assert!(
             store.0.async_support(),
             "cannot use `new_async` without enabling async support on the config"
         );
-        #[cfg(feature = "component-model-async")]
-        {
-            crate::component::concurrent::on_fiber(store, move |store| {
-                Table::_new(store.0, ty, init)
-            })
+        store
+            .on_fiber(|store| Table::_new(store.0, ty, init))
             .await?
-        }
-        #[cfg(not(feature = "component-model-async"))]
-        {
-            let mut store = store;
-            store
-                .on_fiber(|store| Table::_new(store.0, ty, init))
-                .await?
-        }
     }
 
     fn _new(store: &mut StoreOpaque, ty: TableType, init: Ref) -> Result<Table> {
@@ -308,25 +297,14 @@ impl Table {
         delta: u64,
         init: Ref,
     ) -> Result<u64> {
-        let store = store.as_context_mut();
+        let mut store = store.as_context_mut();
         assert!(
             store.0.async_support(),
             "cannot use `grow_async` without enabling async support on the config"
         );
-        #[cfg(feature = "component-model-async")]
-        {
-            crate::component::concurrent::on_fiber(store, move |store| {
-                self.grow(store, delta, init)
-            })
+        store
+            .on_fiber(|store| self.grow(store, delta, init))
             .await?
-        }
-        #[cfg(not(feature = "component-model-async"))]
-        {
-            let mut store = store;
-            store
-                .on_fiber(|store| self.grow(store, delta, init))
-                .await?
-        }
     }
 
     /// Copy `len` elements from `src_table[src_index..]` into
