+++ crates/wasmtime/src/runtime/linker.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "async")]
+use crate::fiber::AsyncCx;
 use crate::func::HostFunc;
 use crate::hash_map::{Entry, HashMap};
 use crate::instance::InstancePre;
@@ -471,28 +473,11 @@ impl<T> Linker<T> {
         );
         assert!(ty.comes_from_same_engine(self.engine()));
         self.func_new(module, name, ty, move |mut caller, params, results| {
-            #[cfg(feature = "component-model-async")]
-            {
-                let async_cx = crate::component::concurrent::AsyncCx::new(&mut caller.store.0);
-                let mut future = Pin::from(func(caller, params, results));
-                match unsafe { async_cx.block_on(future.as_mut()) } {
-                    Ok(Ok(())) => Ok(()),
-                    Ok(Err(trap)) | Err(trap) => Err(trap),
-                }
-            }
-            #[cfg(not(feature = "component-model-async"))]
-            {
-                let async_cx = caller
-                    .store
-                    .as_context_mut()
-                    .0
-                    .async_cx()
-                    .expect("Attempt to spawn new function on dying fiber");
-                let future = func(caller, params, results);
-                match unsafe { async_cx.block_on(Pin::from(future)) } {
-                    Ok(Ok(())) => Ok(()),
-                    Ok(Err(trap)) | Err(trap) => Err(trap),
-                }
+            let async_cx = AsyncCx::new(&mut caller.store.0);
+            let mut future = Pin::from(func(caller, params, results));
+            match async_cx.block_on(future.as_mut()) {
+                Ok(Ok(())) => Ok(()),
+                Ok(Err(trap)) | Err(trap) => Err(trap),
             }
         })
     }
@@ -590,29 +575,12 @@ impl<T> Linker<T> {
         let func = HostFunc::wrap_inner(
             &self.engine,
             move |mut caller: Caller<'_, T>, args: Params| {
-                #[cfg(feature = "component-model-async")]
-                {
-                    let async_cx = crate::component::concurrent::AsyncCx::new(&mut caller.store.0);
-                    let mut future = Pin::from(func(caller, args));
+                let async_cx = AsyncCx::new(&mut caller.store.0);
+                let mut future = Pin::from(func(caller, args));
 
-                    match unsafe { async_cx.block_on(future.as_mut()) } {
-                        Ok(ret) => ret.into_fallible(),
-                        Err(e) => Args::fallible_from_error(e),
-                    }
-                }
-                #[cfg(not(feature = "component-model-async"))]
-                {
-                    let async_cx = caller
-                        .store
-                        .as_context_mut()
-                        .0
-                        .async_cx()
-                        .expect("Attempt to start async function on dying fiber");
-                    let future = func(caller, args);
-                    match unsafe { async_cx.block_on(Pin::from(future)) } {
-                        Ok(ret) => ret.into_fallible(),
-                        Err(e) => Args::fallible_from_error(e),
-                    }
+                match async_cx.block_on(future.as_mut()) {
+                    Ok(ret) => ret.into_fallible(),
+                    Err(e) => Args::fallible_from_error(e),
                 }
             },
         );
