+++ crates/wasmtime/src/runtime/store/async_.rs
@@ -1,16 +1,13 @@
 #[cfg(feature = "call-hook")]
 use crate::CallHook;
+use crate::fiber::{self, AsyncCx};
 use crate::prelude::*;
-use crate::runtime::vm::mpk::{self, ProtectionMask};
-use crate::store::{ResourceLimiterInner, StoreInner, StoreOpaque};
-use crate::{Engine, Store, StoreContextMut, UpdateDeadline};
+use crate::store::{ResourceLimiterInner, StoreInner, StoreOpaque, StoreToken};
+use crate::{AsContextMut, Store, StoreContextMut, UpdateDeadline};
 use anyhow::Context as _;
-use core::cell::UnsafeCell;
 use core::future::Future;
 use core::ops::Range;
-use core::pin::{Pin, pin};
-use core::ptr;
-use core::task::{Context, Poll};
+use core::pin::Pin;
 
 /// An object that can take callbacks when the runtime enters or exits hostcalls.
 #[cfg(feature = "call-hook")]
@@ -19,47 +16,6 @@ pub trait CallHookHandler<T>: Send {
     /// A callback to run when wasmtime is about to enter a host call, or when about to
     /// exit the hostcall.
     async fn handle_call_event(&self, t: StoreContextMut<'_, T>, ch: CallHook) -> Result<()>;
-}
-
-pub struct AsyncState {
-    current_suspend: UnsafeCell<*mut wasmtime_fiber::Suspend<Result<()>, (), Result<()>>>,
-    current_poll_cx: UnsafeCell<PollContext>,
-    /// The last fiber stack that was in use by this store.
-    last_fiber_stack: Option<wasmtime_fiber::FiberStack>,
-}
-
-impl Default for AsyncState {
-    fn default() -> AsyncState {
-        AsyncState {
-            current_suspend: UnsafeCell::new(ptr::null_mut()),
-            current_poll_cx: UnsafeCell::new(PollContext::default()),
-            last_fiber_stack: None,
-        }
-    }
-}
-
-// Lots of pesky unsafe cells and pointers in this structure. This means we need
-// to declare explicitly that we use this in a threadsafe fashion.
-unsafe impl Send for AsyncState {}
-unsafe impl Sync for AsyncState {}
-
-#[derive(Clone, Copy)]
-struct PollContext {
-    future_context: *mut Context<'static>,
-    #[cfg_attr(feature = "component-model-async", allow(dead_code))]
-    guard_range_start: *mut u8,
-    #[cfg_attr(feature = "component-model-async", allow(dead_code))]
-    guard_range_end: *mut u8,
-}
-
-impl Default for PollContext {
-    fn default() -> PollContext {
-        PollContext {
-            future_context: core::ptr::null_mut(),
-            guard_range_start: core::ptr::null_mut(),
-            guard_range_end: core::ptr::null_mut(),
-        }
-    }
 }
 
 impl<T> Store<T> {
@@ -228,24 +184,11 @@ impl<T> StoreInner<T> {
         &'a mut self,
         mk_future: impl FnOnce(&'a mut Self) -> F,
     ) -> Result<F::Output> {
-        #[cfg(feature = "component-model-async")]
-        {
-            let async_cx = crate::component::concurrent::AsyncCx::try_new(self)
-                .context("couldn't create AsyncCx to block on async operation")?;
-            let future = mk_future(self);
-            let mut future = core::pin::pin!(future);
-            unsafe { Ok(async_cx.block_on(future.as_mut())?) }
-        }
-        #[cfg(not(feature = "component-model-async"))]
-        {
-            let cx = self
-                .inner
-                .async_cx()
-                .context("async_cx is not present to block on async operation")?;
-            let future = mk_future(self);
-            let mut future = core::pin::pin!(future);
-            unsafe { cx.block_on(future.as_mut()) }
-        }
+        let async_cx = AsyncCx::try_new(self)
+            .context("couldn't create AsyncCx to block on async operation")?;
+        let future = mk_future(self);
+        let mut future = core::pin::pin!(future);
+        Ok(async_cx.block_on(future.as_mut())?)
     }
 }
 
@@ -256,368 +199,11 @@ impl StoreOpaque {
     /// This function will convert the synchronous `func` into an asynchronous
     /// future. This is done by running `func` in a fiber on a separate native
     /// stack which can be suspended and resumed from.
-    ///
-    /// Most of the nitty-gritty here is how we juggle the various contexts
-    /// necessary to suspend the fiber later on and poll sub-futures. It's hoped
-    /// that the various comments are illuminating as to what's going on here.
-    pub(crate) async fn on_fiber<R>(
+    pub(crate) async fn on_fiber<R: Send + 'static>(
         &mut self,
         func: impl FnOnce(&mut Self) -> R + Send,
     ) -> Result<R> {
-        let config = self.engine().config();
-        debug_assert!(self.async_support());
-        debug_assert!(config.async_stack_size > 0);
-
-        let mut slot = None;
-        let mut future = {
-            let current_poll_cx = self.async_state.current_poll_cx.get();
-            let current_suspend = self.async_state.current_suspend.get();
-            let stack = self.allocate_fiber_stack()?;
-            let track_pkey_context_switch = self.pkey.is_some();
-
-            let engine = self.engine().clone();
-            let slot = &mut slot;
-            let this = &mut *self;
-            let fiber = wasmtime_fiber::Fiber::new(stack, move |keep_going, suspend| {
-                // First check and see if we were interrupted/dropped, and only
-                // continue if we haven't been.
-                keep_going?;
-
-                // Configure our store's suspension context for the rest of the
-                // execution of this fiber. Note that a raw pointer is stored here
-                // which is only valid for the duration of this closure.
-                // Consequently we at least replace it with the previous value when
-                // we're done. This reset is also required for correctness because
-                // otherwise our value will overwrite another active fiber's value.
-                // There should be a test that segfaults in `async_functions.rs` if
-                // this `Replace` is removed.
-                unsafe {
-                    let _reset = Reset(current_suspend, *current_suspend);
-                    *current_suspend = suspend;
-
-                    *slot = Some(func(this));
-                    Ok(())
-                }
-            })?;
-
-            // Once we have the fiber representing our synchronous computation, we
-            // wrap that in a custom future implementation which does the
-            // translation from the future protocol to our fiber API.
-            FiberFuture {
-                fiber: Some(fiber),
-                current_poll_cx,
-                engine,
-                fiber_resume_state: Some(FiberResumeState {
-                    tls: crate::runtime::vm::AsyncWasmCallState::new(),
-                    mpk: if track_pkey_context_switch {
-                        Some(ProtectionMask::all())
-                    } else {
-                        None
-                    },
-                }),
-            }
-        };
-        (&mut future).await?;
-        let stack = future.fiber.take().map(|f| f.into_stack());
-        drop(future);
-        if let Some(stack) = stack {
-            self.deallocate_fiber_stack(stack);
-        }
-
-        return Ok(slot.unwrap());
-
-        struct FiberFuture<'a> {
-            fiber: Option<wasmtime_fiber::Fiber<'a, Result<()>, (), Result<()>>>,
-            current_poll_cx: *mut PollContext,
-            engine: Engine,
-            // See comments in `FiberResumeState` for this
-            fiber_resume_state: Option<FiberResumeState>,
-        }
-
-        // This is surely the most dangerous `unsafe impl Send` in the entire
-        // crate. There are two members in `FiberFuture` which cause it to not
-        // be `Send`. One is `current_poll_cx` and is entirely uninteresting.
-        // This is just used to manage `Context` pointers across `await` points
-        // in the future, and requires raw pointers to get it to happen easily.
-        // Nothing too weird about the `Send`-ness, values aren't actually
-        // crossing threads.
-        //
-        // The really interesting piece is `fiber`. Now the "fiber" here is
-        // actual honest-to-god Rust code which we're moving around. What we're
-        // doing is the equivalent of moving our thread's stack to another OS
-        // thread. Turns out we, in general, have no idea what's on the stack
-        // and would generally have no way to verify that this is actually safe
-        // to do!
-        //
-        // Thankfully, though, Wasmtime has the power. Without being glib it's
-        // actually worth examining what's on the stack. It's unfortunately not
-        // super-local to this function itself. Our closure to `Fiber::new` runs
-        // `func`, which is given to us from the outside. Thankfully, though, we
-        // have tight control over this. Usage of `on_fiber` is typically done
-        // *just* before entering WebAssembly itself, so we'll have a few stack
-        // frames of Rust code (all in Wasmtime itself) before we enter wasm.
-        //
-        // Once we've entered wasm, well then we have a whole bunch of wasm
-        // frames on the stack. We've got this nifty thing called Cranelift,
-        // though, which allows us to also have complete control over everything
-        // on the stack!
-        //
-        // Finally, when wasm switches back to the fiber's starting pointer
-        // (this future we're returning) then it means wasm has reentered Rust.
-        // Suspension can only happen via the `block_on` function of an
-        // `AsyncCx`. This, conveniently, also happens entirely in Wasmtime
-        // controlled code!
-        //
-        // There's an extremely important point that should be called out here.
-        // User-provided futures **are not on the stack** during suspension
-        // points. This is extremely crucial because we in general cannot reason
-        // about Send/Sync for stack-local variables since rustc doesn't analyze
-        // them at all. With our construction, though, we are guaranteed that
-        // Wasmtime owns all stack frames between the stack of a fiber and when
-        // the fiber suspends (and it could move across threads). At this time
-        // the only user-provided piece of data on the stack is the future
-        // itself given to us. Lo-and-behold as you might notice the future is
-        // required to be `Send`!
-        //
-        // What this all boils down to is that we, as the authors of Wasmtime,
-        // need to be extremely careful that on the async fiber stack we only
-        // store Send things. For example we can't start using `Rc` willy nilly
-        // by accident and leave a copy in TLS somewhere. (similarly we have to
-        // be ready for TLS to change while we're executing wasm code between
-        // suspension points).
-        //
-        // While somewhat onerous it shouldn't be too too hard (the TLS bit is
-        // the hardest bit so far). This does mean, though, that no user should
-        // ever have to worry about the `Send`-ness of Wasmtime. If rustc says
-        // it's ok, then it's ok.
-        //
-        // With all that in mind we unsafely assert here that wasmtime is
-        // correct. We declare the fiber as only containing Send data on its
-        // stack, despite not knowing for sure at compile time that this is
-        // correct. That's what `unsafe` in Rust is all about, though, right?
-        unsafe impl Send for FiberFuture<'_> {}
-
-        impl FiberFuture<'_> {
-            fn fiber(&self) -> &wasmtime_fiber::Fiber<'_, Result<()>, (), Result<()>> {
-                self.fiber.as_ref().unwrap()
-            }
-
-            /// This is a helper function to call `resume` on the underlying
-            /// fiber while correctly managing Wasmtime's state that the fiber
-            /// may clobber.
-            ///
-            /// ## Return Value
-            ///
-            /// * `Ok(Ok(()))` - the fiber successfully completed and yielded a
-            ///    successful result.
-            /// * `Ok(Err(e))` - the fiber successfully completed and yielded
-            ///   an error as a result of computation.
-            /// * `Err(())` - the fiber has not finished and it is suspended.
-            fn resume(&mut self, val: Result<()>) -> Result<Result<()>, ()> {
-                unsafe {
-                    let prev = self.fiber_resume_state.take().unwrap().replace();
-                    let restore = Restore {
-                        fiber: self,
-                        prior_fiber_state: Some(prev),
-                    };
-                    return restore.fiber.fiber().resume(val);
-                }
-
-                struct Restore<'a, 'b> {
-                    fiber: &'a mut FiberFuture<'b>,
-                    prior_fiber_state: Option<PriorFiberResumeState>,
-                }
-
-                impl Drop for Restore<'_, '_> {
-                    fn drop(&mut self) {
-                        unsafe {
-                            self.fiber.fiber_resume_state =
-                                Some(self.prior_fiber_state.take().unwrap().replace());
-                        }
-                    }
-                }
-            }
-        }
-
-        impl Future for FiberFuture<'_> {
-            type Output = Result<()>;
-
-            fn poll(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-                // We need to carry over this `cx` into our fiber's runtime
-                // for when it tries to poll sub-futures that are created. Doing
-                // this must be done unsafely, however, since `cx` is only alive
-                // for this one singular function call. Here we do a `transmute`
-                // to extend the lifetime of `Context` so it can be stored in
-                // our `Store`, and then we replace the current polling context
-                // with this one.
-                //
-                // Note that the replace is done for weird situations where
-                // futures might be switching contexts and there's multiple
-                // wasmtime futures in a chain of futures.
-                //
-                // On exit from this function, though, we reset the polling
-                // context back to what it was to signify that `Store` no longer
-                // has access to this pointer.
-                let guard = self
-                    .fiber()
-                    .stack()
-                    .guard_range()
-                    .unwrap_or(core::ptr::null_mut()..core::ptr::null_mut());
-                unsafe {
-                    let _reset = Reset(self.current_poll_cx, *self.current_poll_cx);
-                    *self.current_poll_cx = PollContext {
-                        future_context: core::mem::transmute::<
-                            &mut Context<'_>,
-                            *mut Context<'static>,
-                        >(cx),
-                        guard_range_start: guard.start,
-                        guard_range_end: guard.end,
-                    };
-
-                    // After that's set up we resume execution of the fiber, which
-                    // may also start the fiber for the first time. This either
-                    // returns `Ok` saying the fiber finished (yay!) or it
-                    // returns `Err` with the payload passed to `suspend`, which
-                    // in our case is `()`.
-                    match self.resume(Ok(())) {
-                        Ok(result) => Poll::Ready(result),
-
-                        // If `Err` is returned that means the fiber polled a
-                        // future but it said "Pending", so we propagate that
-                        // here.
-                        //
-                        // An additional safety check is performed when leaving
-                        // this function to help bolster the guarantees of
-                        // `unsafe impl Send` above. Notably this future may get
-                        // re-polled on a different thread. Wasmtime's
-                        // thread-local state points to the stack, however,
-                        // meaning that it would be incorrect to leave a pointer
-                        // in TLS when this function returns. This function
-                        // performs a runtime assert to verify that this is the
-                        // case, notably that the one TLS pointer Wasmtime uses
-                        // is not pointing anywhere within the stack. If it is
-                        // then that's a bug indicating that TLS management in
-                        // Wasmtime is incorrect.
-                        Err(()) => {
-                            if let Some(range) = self.fiber().stack().range() {
-                                crate::runtime::vm::AsyncWasmCallState::assert_current_state_not_in_range(range);
-                            }
-                            Poll::Pending
-                        }
-                    }
-                }
-            }
-        }
-
-        // Dropping futures is pretty special in that it means the future has
-        // been requested to be cancelled. Here we run the risk of dropping an
-        // in-progress fiber, and if we were to do nothing then the fiber would
-        // leak all its owned stack resources.
-        //
-        // To handle this we implement `Drop` here and, if the fiber isn't done,
-        // resume execution of the fiber saying "hey please stop you're
-        // interrupted". Our `Trap` created here (which has the stack trace
-        // of whomever dropped us) will then get propagated in whatever called
-        // `block_on`, and the idea is that the trap propagates all the way back
-        // up to the original fiber start, finishing execution.
-        //
-        // We don't actually care about the fiber's return value here (no one's
-        // around to look at it), we just assert the fiber finished to
-        // completion.
-        impl Drop for FiberFuture<'_> {
-            fn drop(&mut self) {
-                if self.fiber.is_none() {
-                    return;
-                }
-
-                if !self.fiber().done() {
-                    let result = self.resume(Err(anyhow!("future dropped")));
-                    // This resumption with an error should always complete the
-                    // fiber. While it's technically possible for host code to
-                    // catch the trap and re-resume, we'd ideally like to
-                    // signal that to callers that they shouldn't be doing
-                    // that.
-                    debug_assert!(result.is_ok());
-
-                    // Note that `result` is `Ok(r)` where `r` is either
-                    // `Ok(())` or `Err(e)`. If it's an error that's disposed of
-                    // here. It's expected to be a propagation of the `future
-                    // dropped` error created above.
-                }
-
-                self.fiber_resume_state.take().unwrap().dispose();
-
-                unsafe {
-                    self.engine
-                        .allocator()
-                        .deallocate_fiber_stack(self.fiber.take().unwrap().into_stack());
-                }
-            }
-        }
-
-        /// State of the world when a fiber last suspended.
-        ///
-        /// This structure represents global state that a fiber clobbers during
-        /// its execution. For example TLS variables are updated, system
-        /// resources like MPK masks are updated, etc. The purpose of this
-        /// structure is to track all of this state and appropriately
-        /// save/restore it around fiber suspension points.
-        struct FiberResumeState {
-            /// Saved list of `CallThreadState` activations that are stored on a
-            /// fiber stack.
-            ///
-            /// This is a linked list that references stack-stored nodes on the
-            /// fiber stack that is currently suspended. The
-            /// `AsyncWasmCallState` type documents this more thoroughly but the
-            /// general gist is that when we this fiber is resumed this linked
-            /// list needs to be pushed on to the current thread's linked list
-            /// of activations.
-            tls: crate::runtime::vm::AsyncWasmCallState,
-
-            /// Saved MPK protection mask, if enabled.
-            ///
-            /// When MPK is enabled then executing WebAssembly will modify the
-            /// processor's current mask of addressable protection keys. This
-            /// means that our current state may get clobbered when a fiber
-            /// suspends. To ensure that this function preserves context it
-            /// will, when MPK is enabled, save the current mask when this
-            /// function is called and then restore the mask when the function
-            /// returns (aka the fiber suspends).
-            mpk: Option<ProtectionMask>,
-        }
-
-        impl FiberResumeState {
-            unsafe fn replace(self) -> PriorFiberResumeState {
-                let tls = self.tls.push();
-                let mpk = swap_mpk_states(self.mpk);
-                PriorFiberResumeState { tls, mpk }
-            }
-
-            fn dispose(self) {
-                self.tls.assert_null();
-            }
-        }
-
-        struct PriorFiberResumeState {
-            tls: crate::runtime::vm::PreviousAsyncWasmCallState,
-            mpk: Option<ProtectionMask>,
-        }
-
-        impl PriorFiberResumeState {
-            unsafe fn replace(self) -> FiberResumeState {
-                let tls = self.tls.restore();
-                let mpk = swap_mpk_states(self.mpk);
-                FiberResumeState { tls, mpk }
-            }
-        }
-
-        fn swap_mpk_states(mask: Option<ProtectionMask>) -> Option<ProtectionMask> {
-            mask.map(|mask| {
-                let current = mpk::current_mask();
-                mpk::allow(mask);
-                current
-            })
-        }
+        fiber::on_fiber(self, func).await
     }
 
     #[cfg(feature = "gc")]
@@ -678,30 +264,6 @@ impl StoreOpaque {
         log::trace!("End trace GC roots")
     }
 
-    /// Yields the async context, assuming that we are executing on a fiber and
-    /// that fiber is not in the process of dying. This function will return
-    /// None in the latter case (the fiber is dying), and panic if
-    /// `async_support()` is false.
-    #[inline]
-    pub fn async_cx(&self) -> Option<AsyncCx> {
-        assert!(self.async_support());
-
-        let poll_cx_box_ptr = self.async_state.current_poll_cx.get();
-        if poll_cx_box_ptr.is_null() {
-            return None;
-        }
-
-        let poll_cx_inner_ptr = unsafe { *poll_cx_box_ptr };
-        if poll_cx_inner_ptr.future_context.is_null() {
-            return None;
-        }
-
-        Some(AsyncCx {
-            current_suspend: self.async_state.current_suspend.get(),
-            current_poll_cx: unsafe { &raw mut (*poll_cx_box_ptr).future_context },
-        })
-    }
-
     /// Yields execution to the caller on out-of-gas or epoch interruption.
     ///
     /// This only works on async futures and stores, and assumes that we're
@@ -718,21 +280,23 @@ impl StoreOpaque {
         // to clean up this fiber. Do so by raising a trap which will
         // abort all wasm and get caught on the other side to clean
         // things up.
-        unsafe {
-            self.async_cx()
-                .expect("attempted to pull async context during shutdown")
-                .block_on(Pin::new_unchecked(&mut future))
-        }
+        AsyncCx::try_new(self)
+            .expect("attempted to pull async context during shutdown")
+            .block_on(unsafe { Pin::new_unchecked(&mut future) })
     }
 
-    fn allocate_fiber_stack(&mut self) -> Result<wasmtime_fiber::FiberStack> {
+    pub(crate) fn async_guard_range(&mut self) -> Range<*mut u8> {
+        unsafe { (*self.async_state()).async_guard_range() }
+    }
+
+    pub(crate) fn allocate_fiber_stack(&mut self) -> Result<wasmtime_fiber::FiberStack> {
         if let Some(stack) = self.async_state.last_fiber_stack.take() {
             return Ok(stack);
         }
         self.engine().allocator().allocate_fiber_stack()
     }
 
-    fn deallocate_fiber_stack(&mut self, stack: wasmtime_fiber::FiberStack) {
+    pub(crate) fn deallocate_fiber_stack(&mut self, stack: wasmtime_fiber::FiberStack) {
         self.flush_fiber_stack();
         self.async_state.last_fiber_stack = Some(stack);
     }
@@ -746,113 +310,20 @@ impl StoreOpaque {
             }
         }
     }
-
-    pub(crate) fn async_guard_range(&mut self) -> Range<*mut u8> {
-        #[cfg(feature = "component-model-async")]
-        {
-            unsafe { (*self.concurrent_async_state()).async_guard_range() }
-        }
-        #[cfg(not(feature = "component-model-async"))]
-        unsafe {
-            let ptr = self.async_state.current_poll_cx.get();
-            (*ptr).guard_range_start..(*ptr).guard_range_end
-        }
-    }
 }
 
 impl<T> StoreContextMut<'_, T> {
     /// Executes a synchronous computation `func` asynchronously on a new fiber.
-    pub(crate) async fn on_fiber<R>(
+    pub(crate) async fn on_fiber<R: Send + 'static>(
         &mut self,
         func: impl FnOnce(&mut StoreContextMut<'_, T>) -> R + Send,
     ) -> Result<R>
     where
         T: Send + 'static,
     {
+        let token = StoreToken::new(self.as_context_mut());
         self.0
-            .on_fiber(|opaque| {
-                let store = unsafe { opaque.traitobj().cast::<StoreInner<T>>().as_mut() };
-                func(&mut StoreContextMut(store))
-            })
+            .on_fiber(|opaque| func(&mut token.as_context_mut(opaque.traitobj_mut())))
             .await
-    }
-}
-
-pub struct AsyncCx {
-    current_suspend: *mut *mut wasmtime_fiber::Suspend<Result<()>, (), Result<()>>,
-    current_poll_cx: *mut *mut Context<'static>,
-}
-
-impl AsyncCx {
-    /// Blocks on the asynchronous computation represented by `future` and
-    /// produces the result here, in-line.
-    ///
-    /// This function is designed to only work when it's currently executing on
-    /// a native fiber. This fiber provides the ability for us to handle the
-    /// future's `Pending` state as "jump back to whomever called the fiber in
-    /// an asynchronous fashion and propagate `Pending`". This tight coupling
-    /// with `on_fiber` below is what powers the asynchronicity of calling wasm.
-    /// Note that the asynchronous part only applies to host functions, wasm
-    /// itself never really does anything asynchronous at this time.
-    ///
-    /// This function takes a `future` and will (appear to) synchronously wait
-    /// on the result. While this function is executing it will fiber switch
-    /// to-and-from the original frame calling `on_fiber` which should be a
-    /// guarantee due to how async stores are configured.
-    ///
-    /// The return value here is either the output of the future `T`, or a trap
-    /// which represents that the asynchronous computation was cancelled. It is
-    /// not recommended to catch the trap and try to keep executing wasm, so
-    /// we've tried to liberally document this.
-    pub unsafe fn block_on<F>(&self, future: F) -> Result<F::Output>
-    where
-        F: Future + Send,
-    {
-        let mut future = pin!(future);
-
-        // Take our current `Suspend` context which was configured as soon as
-        // our fiber started. Note that we must load it at the front here and
-        // save it on our stack frame. While we're polling the future other
-        // fibers may be started for recursive computations, and the current
-        // suspend context is only preserved at the edges of the fiber, not
-        // during the fiber itself.
-        //
-        // For a little bit of extra safety we also replace the current value
-        // with null to try to catch any accidental bugs on our part early.
-        // This is all pretty unsafe so we're trying to be careful...
-        //
-        // Note that there should be a segfaulting test  in `async_functions.rs`
-        // if this `Reset` is removed.
-        let suspend = *self.current_suspend;
-        let _reset = Reset(self.current_suspend, suspend);
-        *self.current_suspend = ptr::null_mut();
-        assert!(!suspend.is_null());
-
-        loop {
-            let future_result = {
-                let poll_cx = *self.current_poll_cx;
-                let _reset = Reset(self.current_poll_cx, poll_cx);
-                *self.current_poll_cx = ptr::null_mut();
-                assert!(!poll_cx.is_null());
-                future.as_mut().poll(&mut *poll_cx)
-            };
-
-            match future_result {
-                Poll::Ready(t) => break Ok(t),
-                Poll::Pending => {}
-            }
-
-            (*suspend).suspend(())?;
-        }
-    }
-}
-
-struct Reset<T: Copy>(*mut T, T);
-
-impl<T: Copy> Drop for Reset<T> {
-    fn drop(&mut self) {
-        unsafe {
-            *self.0 = self.1;
-        }
     }
 }
